requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.3
django-toolbelt==0.0.1
gunicorn==19.1.1
oauthlib==0.7.2
psycopg2==2.5.4
PyJWT==0.4.1
python-social-auth==0.2.1
python3-openid==3.0.5
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
static3==0.5.1
